Charge species must be a float
Must be merged after #248.
Closes #260.

See merge request gysela-developpers/gyselalibxx!573

--------------------------------------------

+++ simulations/geometryXVx/neutrals/pdi_out_neutrals.yml.hpp
@@ -24,7 +24,7 @@ metadata:
   fdistribu_charges_extents : { type: array, subtype: int64, size: 1 }
   fdistribu_charges:
     type: array
-    subtype: int
+    subtype: double
     size: [ '$fdistribu_charges_extents[0]' ]
   fdistribu_masses_extents : { type: array, subtype: int64, size: 1 }
   fdistribu_masses:
